requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.5
django-filter==22.1
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.4
djangorestframework==3.14.0
graphene==3.2.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
promise==2.3
psycopg2==2.9.5
PyJWT==2.6.0
pytz==2022.7
Rx==1.6.3
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
